printf: review 1
