+++ src/kernel/mem_utils.h
@@ -1,0 +1,3 @@
+void* memcpy(void* dst, const void* src, int n);
+
+void* memzero(void* dst, int n);
